move menu bar handlers to a separate file
