+++ js/keybindings.js
@@ -16,117 +16,9 @@ var browserUI = require('browserUI.js')
 var focusMode = require('focusMode.js')
 var urlParser = require('util/urlParser.js')
 var settings = require('util/settings/settings.js')
+var findinpage = require('findinpage.js')
 
 var keyMap
-
-ipc.on('zoomIn', function () {
-  webviewGestures.zoomWebviewIn(tabs.getSelected())
-})
-
-ipc.on('zoomOut', function () {
-  webviewGestures.zoomWebviewOut(tabs.getSelected())
-})
-
-ipc.on('zoomReset', function () {
-  webviewGestures.resetWebviewZoom(tabs.getSelected())
-})
-
-ipc.on('print', function () {
-  if (PDFViewer.isPDFViewer(tabs.getSelected())) {
-    PDFViewer.printPDF(tabs.getSelected())
-  } else if (readerView.isReader(tabs.getSelected())) {
-    readerView.printArticle(tabs.getSelected())
-  } else {
-    // TODO figure out why webContents.print() doesn't work in Electron 4
-    webviews.get(tabs.getSelected()).executeJavaScript('window.print()')
-  }
-})
-
-ipc.on('findInPage', function () {
-  findinpage.start()
-})
-
-ipc.on('inspectPage', function () {
-  webviews.get(tabs.getSelected()).openDevTools()
-})
-
-ipc.on('showReadingList', function () {
-  // open the searchbar with "!readinglist " as the input
-  tabBar.enterEditMode(tabs.getSelected(), '!readinglist ')
-})
-
-ipc.on('showBookmarks', function () {
-  tabBar.enterEditMode(tabs.getSelected(), '!bookmarks ')
-})
-
-ipc.on('showHistory', function () {
-  tabBar.enterEditMode(tabs.getSelected(), '!history ')
-})
-
-ipc.on('duplicateTab', function (e) {
-  /* new tabs can't be created in focus mode */
-  if (focusMode.enabled()) {
-    focusMode.warn()
-    return
-  }
-
-  // can't duplicate if tabs is empty
-  if (tabs.isEmpty()) {
-    return
-  }
-
-  const sourceTab = tabs.get(tabs.getSelected())
-  // strip tab id so that a new one is generated
-  const newTab = tabs.add({...sourceTab, id: undefined})
-
-  browserUI.addTab(newTab, { enterEditMode: false })
-})
-
-ipc.on('addTab', function (e, data) {
-  /* new tabs can't be created in focus mode */
-  if (focusMode.enabled()) {
-    focusMode.warn()
-    return
-  }
-
-  // if opening a URL (instead of adding an empty tab), and the current tab is empty, navigate the current tab rather than creating another one
-  if (!tabs.get(tabs.getSelected()).url && data.url) {
-    browserUI.navigate(tabs.getSelected(), data.url)
-  } else {
-    var newTab = tabs.add({
-      url: data.url || ''
-    })
-
-    browserUI.addTab(newTab, {
-      enterEditMode: !data.url // only enter edit mode if the new tab is empty
-    })
-  }
-})
-
-ipc.on('saveCurrentPage', function () {
-  var currentTab = tabs.get(tabs.getSelected())
-
-  // new tabs cannot be saved
-  if (!currentTab.url) {
-    return
-  }
-
-  // if the current tab is a PDF, let the PDF viewer handle saving the document
-  if (PDFViewer.isPDFViewer(tabs.getSelected())) {
-    PDFViewer.savePDF(tabs.getSelected())
-    return
-  }
-
-  var savePath = remote.dialog.showSaveDialog(remote.getCurrentWindow(), {})
-
-  // savePath will be undefined if the save dialog is canceled
-  if (savePath) {
-    if (!savePath.endsWith('.html')) {
-      savePath = savePath + '.html'
-    }
-    webviews.get(currentTab.id).savePage(savePath, 'HTMLComplete', function () {})
-  }
-})
 
 function addPrivateTab () {
   /* new tabs can't be created in focus mode */
@@ -144,35 +36,6 @@ function addPrivateTab () {
   })
   browserUI.addTab(privateTab)
 }
-
-ipc.on('addPrivateTab', addPrivateTab)
-
-ipc.on('addTask', function () {
-  /* new tasks can't be created in focus mode */
-  if (focusMode.enabled()) {
-    focusMode.warn()
-    return
-  }
-
-  browserUI.addTask()
-  taskOverlay.show()
-  setTimeout(function () {
-    taskOverlay.hide()
-    tabBar.enterEditMode(tabs.getSelected())
-  }, 600)
-})
-
-ipc.on('goBack', function () {
-  try {
-    webviews.get(tabs.getSelected()).goBack()
-  } catch (e) {}
-})
-
-ipc.on('goForward', function () {
-  try {
-    webviews.get(tabs.getSelected()).goForward()
-  } catch (e) {}
-})
 
 var menuBarShortcuts = ['mod+t', 'shift+mod+p', 'mod+n'] // shortcuts that are already used for menu bar items
